Renamed TranslationManager file name
